+++ app/(main)/(routes)/servers/[serverId]/channels/[channelId]/page.tsx
@@ -118,6 +118,7 @@ const ChannelIdPage = async ({
                   // serverId: params.serverId,
                   serverId: server.id,
                 }}
+                memberListProp={members}
               />
             </>
             )}